refactor(LottoGame): Utils 성 클래스 및 WinningNumbers 삭제
