+++ src/test/java/lotto/domain/LottoNumberGeneratorTest.java
@@ -1,5 +1,6 @@
-package lotto.utils;
+package lotto.domain;
 
+import lotto.domain.LottoNumberGenerator;
 import lotto.domain.Lottoes;
 import lotto.domain.PurchaseAmount;
 import org.junit.jupiter.api.DisplayName;
